Replace 'gitignore' crate with 'ignore'
This commit substitutes the 'gitignore' crate with the 'ignore' crate for enhanced performance and flexibility. The change includes updates to the relevant configuration files and code adjustments to utilize the new crate's API. This switch was motivated by the 'ignore' crate's more efficient file filtering capabilities, which align better  the project's needs. Additionally, it was necessary to make this switch as the 'gitignore' crate is no longer actively maintained, and its author has discouraged its further use.